Refactor 675. Solution still not worked.
